Remove references to GL* types
As a first step to removing dependence on OpenGL of the core IceT,
remove all the GL* types (i.e. GLint, GLboolean, GLfloat, etc.).
This removes most of the dependence on OpenGL right there.  (There
are only a limited amount of OpenGL function calls left.

As part of this process I have also moved the basic header files out
of the include/GL directory and into the base include directory.
I also changed the names from ice-t to IceT to better conform to the
current name convention.