Change IceTPointerArithmetic to signed.
The IceTPointerArithmetic type (and the subsequently derived
IcetSizeType) were declared as unsigned integers of the correct
length.  The problem is that it became odd to represent negative
numbers, which can happen when doing actual arithmetic.  I
recently traced down a bug in icetGetViewportProject where
an IceTInt was subtracted by an IceTSizeType (resulting in a
negative number) and then cast to an IceTDouble.  Because the
twos complement was interpreted as a positive number we got a
huge number that was totally unexpected.  This change should
prevent mistakes like these.