Made macro header for compress "switch."
The compress_func_body.h that allowed us to reuse the code for
multiple forms of compress (sort of like generic template programming
in C++) still required the user code to create a bunch of cases
for the different type of image formats.  To consolidate that too,
renamed compress_func_body.h to compress_template_body.h.  The new
compress_func_body.h now contains contains all the types of formats
that can be encountered in an image.  The decompress functions should
soon follow.