Replaced input/output buffers with composite mode.
Got rid of the input output state.  This will be determined
by whatever kind of image the callback returns.  Added a
function for getting the maximum needed buffer size for
any image format.  Also added a state variable for the
composite mode.  This makes the compositing explicit.  I
have not changed the compositing methods to use the
explicit variable, and I'm not sure that I'm going to,
but we can and should check images returned that they
conform to the chosen composite mode.