Updated the direct strategy to new image objects.
Also formalized the concept of a "null" image.  This is
convienient for returning an image when no such image
exists.  For example, the strategies can now return the
null image to signify that the display for the current
process is empty.  This is important because we no longer
have the information about what color and depth formats
to use.