Adjusting image for output messing up subsequent iter in serial.
The bswap composite method calls icetImageAdjustForOutput on its
image before collecting the final image data.  This is fine to
return the image, but screws up the serial strategy on the
iteration for the next tile because it reuses this image for the
next input.  To get around this, add icetImageAdjustForInput that
restores the image to the expected input buffers.