Allocated image of wrong size in split strategy.
The outgoing sparse image was allocated with fragment_size pixels.
However, fragment_size has to do with the expected size of the
incoming fragments for the tile we are participating with.  The
fragments we send for other tiles might have a different split and
there fore different fragment sizes.  When sending to a tile with
larger fragments (i.e. less processes compositing them), the
procedure fails because the sparse image buffer is not big enough.