Unify seed in RandomTransform test.
Previously, each process independently created a seed with the
time function.  Easy to implement, but problematic if not all
the time is synchronized.  Instead, create a seed on the root
process and send to the rest.  The rest modify their seed
slightly.  This way you can replicate a failure by just forcing
one seed value rather than worry about all the processes.