Remove OpenGL from draw.c.
All the functions in draw.c now no longer depend on OpenGL at all.
The biggest interface change to this effect is that icetDrawFrame()
now requires the user to pass in the projection/modelview matrices
and the background color.  A new function in the OpenGL layer called
icetGLDrawFrame() behaves like the old function by grabbing these
values from OpenGL.