Readback all tiles in floating viewport render.
In the case of a floating viewport render, the projection was
properly set to include all the geometry.  However the
readback_viewport that was passed to the render callback was
wrong.  It only included the viewport for the current tile
requested.  Thus, the callback did not provide the pixels for
the other tiles even though they were rendered.  The
readback_viewport is now set to include all valid pixels for
all involved tiles.

This problem appeared when running the RandomTransform test on
8 processes with a base seed of 1276806597.