Handle when my fragment truncated in split strategy.
In the split strategy, the last fragment is truncated if the number
of pixels in the image does not divide by the number of processes in
the group.  The message from the rendering process to the compositing
process was properly truncated.  However, the compositing process
did not take into acount truncation and could therefore send a too-big
image to the display process, which in turn caused a buffer overrun.
(In this case, it happened to overrun into the communication request
objects.)

Caught by RandomTransform test running on 8 processes on blackrose
with base seed of 1277230396.

Also made a minor style change to the split code.  No longer reusing
the my_* variables when receiving the displayed image.  Instead,
creating new variables named displayed_*.