Get rid of buckets.
Way back when I first wrote IceT, I included some convienience
functions to implement "buckets," which were also described in
the original paper.  The idea is that you break up your geometry
within a process and in a render callback you only render those
buckets visible in the projection.  That way when you have to
render multiple tiles, you can reduce the number of primitives
that have to be rendered multiple times.

The theory is all sound and still a good idea.  However, I don't
think these convienience functions are all that helpful.  If
your appliation already has a means of breaking up geometry,
then adding a check to see if it lies in the projection is pretty
straightforward.  Thus, I never bothered to document the buckets
code or even write a test for them.  I suspect the only code to
ever use these buckes is languishing in the IceT version control
archive and relies on an I/O library that fell into the bit
bucket years ago.

If I'm wrong and someone uses these functions, I can resurect
them.  But I suspect that I won't.