Destroy inflate texture in callback.
The icetDestroyContext function needs to delete the OpenGL texture
used for inflating images.  This was done directly in the function,
which made it rely on the OpenGL library.  To get around this, the
core layer defines a callback for the destruction of the rendering
layer.  It is in this callback that the OpenGL layer deletes the
texture.

The core layer should now be completely free of OpenGL dependencies.
The next step is to add CMake configure options and appropriate
target links to allow you to compile without OpenGL.