Target link libraries only as necessary.
Previously, the OpenGL and MPI libraries were linked to everything
with a general LINK_LIBRARIES CMake command.  Now, only use
TARGET_LINK_LIBRARIES to bind libraries to a specific target.  Not
only is this cleaner, but it also helps verify that libraries that
should not depend on these libraries don't.