Rename GL layer filenames.
Renamed the GL layer filenames to look like the filenames in the
other libraries.  I actually don't like the names of the other
files, but I am loath to change all of those files and totally
screw up the file history for everything in IceT.  Therefore,
I'll just change these files to conform to the crummy, half-ass
convention.