Make IceTContext a struct pointer.
Previously IceTContext was a typedef'ed integer (that pointed
into an array of context structs).  Now it is a pointer to a
struct whose internals are not exposed.  This maintains the
opaque-ness of the data structure but also introduces type
safety.  It also happens to simplify the implementation.

As part of this, I also hid the structure of the context from
anything but context.c and hid the state value structure from
anything but state.c.