Combine IceTCore and IceTStrategies libraries.
Changed the build system to include all of the strategies into the
IceTCore library.  The separation was mostly a coding convienience
on the library side.  It was rather silly for the user perspective
as you always used both libraries.  Also, this was becoming less
convienient as the core library really relies on things in the
stratgies library and you have to carefully pass pointers and
structures around.

This change is driven by some errors I am encountering that might
be caused by using uninitialized strategy objects.  My next step is
to just identify stratgies by enumerations.  This will get rid of
any such problems.