Change icetImageGetColor/Depth to use short suffixes.
The icetImageGetColor* and icetImageGetDepth* now use the abbriviated
OpenGL convention of using single characters to identify type (such as
ub and f) instead of longer names (such as UByte and Float).  I am
no longer enamored with this convention, but it is used elsewhere in
IceT (e.g. icetBoundingBox), and I am a fan of consistency.