Make all image formats non-zero.
Previously ICET_IMAGE_COLOR_NONE and ICET_IMAGE_DEPTH_NONE were
zero.  They are now 0xC000 and 0xD000, respectively.  The logic
behind this change is to help ensure that the format is always
initialized to something and to prevent format mixup (using the
color identifier for depth and vice versa).