Fixes to automatic k selection.
There were problems selecting k when MAGIC_K was not a factor.  The
first and major problem was that the group was always divided by
MAGIC_K rather than the k actually selected.  Thus, the k factors
did not always multiply to the group size.  The second problem was
that a mistake in the testing of other k values was done incorrectly
and it always picked the total group.

This change fixes those problems and also changes the test to look
both above and below the magic k value.  Thus, it will favor 12 over
2.