Handle empty pieces in binary swap.
To simplify the indexing of image partitions in binary swap, the algorithm
pretends that the image is the next largest factor of the number of
processes.  This ensures that the image will always be divided evenly
and consistently.

As a consiquence, the partition at the end of the image might get
truncated.  This was handled, but was not handled was the possibility
that the last partition was of size 0 and the next largest is truncated.
This can happen if the final partition sizes contain less pixels than
number of processes.  The problem was most easily demonstrated by running
the OddImageSizes test on 8 processes.  (4 processes would probably work
too, but I did not test that.)

This fix should also handle any situation where the size of the image
has fewer pixels than number of processes (which is also the case in
the OddImageSizes test).