Radix-k no longer has MAX_K and MAX_ROUNDS.
When running on large numbers of processes and/or having processes with
large factors, these could cause the algorithm to fail.  The arrays need
for k and rounds are now dynamically allocated.

To facilitate this, I consolidated several arrays into a single array of
a struct containing information pertaining to a particular partner in
a group.  I also broke the algorithm into multiple functions to make it
easier to support and debug.