Use 32-bit integer for IceTSizeType.
Previously the width of IceTSizeType was the same width of a pointer.
That makes sense, but even on a 64-bit system it makes no sense to
create 2 GB buffers.  32-bit sizes and offsets are sufficient for any
buffer to by used by IceT.

That in itself is no big deal.  However, I am shortly going to need
to be able to transfer arrays of IceTSizeType, but MPI has no built in
type for 64-bit integers.  Rather than muck up that code, I changed
the size type.