Single image strategies no longer have to collect.
The calling convention for a single image stragegy is now to return
an offset and size of the local partition, and it is expected that
the results are partitioned amongst processes.  There is now a new
function that must be called to gather the results.

The behavior of the actual single image strategies has not yet
changed.  I will soon remove the collection part of thsese algorithms.