Convert tree composite to composite with sparse images.
The plan is to soon have all single image strategies take and return
sparse images rather than raw images.  This should eventually cut down
on the time to fill the background and then check it again for
compression.  This is a first step in that the tree composite algorithm
internally composites compressed images with other compressed images.
The main purpose for now is to debug the compressed-compressed
composite.