Add memory checks for state variables.
In an attempt to make using state variables more safe, these changes
add some checks for bad memory usage.  For every state variable
allocated, extra memory is added to the beginning and end of the
buffer (in debug mode).  These padding are filled with known values.
When the variable is latter accessed, it is checked to make sure
that these values remain intact (thus there have been no overruns).
Allocated buffers are also cleared to 0xDC to make it obvious when
uninitialized memory is used.