Fix issue with reduce when compositing one tile, displaying another.
This only happened when the display tile was empty.  But what would
happen is that the collect algorithm would prepare the result image
for output and run through the collect operation anyway (even though
it would otherwise have no effect).  If this happened before the
process participated in the collect for the image it composited, then
the image might have the wrong format for uncompression.

This change also conserves one image buffer in reduce.