Black backgrounds in SimpleTiming
Add a flag to SimpleTiming to turn black background on and off.  If
in transparent mode and the background is anything other than black
(with no opacity), then IceT has to "correct" the background by
blending the composite image to the background.  Although important
in practice, it is not worth timing when comparing algorithms.

Now that I think about it, the correction is done completely on the
display node.  Why not have the correction happen right before the
collection so that this time is shared amongst processes?