Split sizes based on final split size.
Changed icetSparseImageSplit and associated functions to base the size
of the splits based on the eventual number of splits done by the end
of the algorithm.  Each of these functions takes a new
eventual_num_partitions argument that is the number of times the
image will be split by the end of the algorithm.  If you partition
the image by factors of this number, you will get the same partition
sizes and order as if you partitioned by that number to begin with.