Add ICET_ADD_LIBRARY macro for special instructions on IceT libs.
All the IceT libraries contain some special instructions for the
builds (such as install targets).  Encapsulate these in a single
macro.

The motivation for this change was to move the compiler warning
flags specifically to the files compiled in IceT libraries (or,
more to the point, outside of the test files).  I don't feel the
test scaffolding needs the same strickness as the libraries.
Also, I need to set an environment variable in one of the tests,
and there is no means of doing that when maintaining ansi
compatibility.