Check size of returned image against valid tile, not display tile.
Previously, the strategy always returned an image for the display
tile, and the calling icetDrawFrame checked the size of that returned
image against the expected size of the display.  With the uncollected
option, any process can return an image for any tile.  Thus, we need
to check against the valid tile rather than the display tile.  If
collection is on, also check to make sure the valid tile is the same
as the display tile.