Clean up memory allocations.
In most places malloc is called, check to make sure that the returned
pointer is not null, and emit an out of memory error if it is.

Found several places where free was not always called for all allocated
memory.

When possible, remove direct calls to malloc in the code and replace
with the state variable memory management.