Suppress warning about assumption of no signed overflow.
The gcc compiler wants to optimize a conditional that is always true,
but only if we assume that there is no signed overflow.  (That is,
there are no negitive numbers.)  It happens to be true wtih the numbers
I give, so do a bunch of casts to tell the compiler that.