When allocating a state buffer, stomp on memory in debug mode.
There was this stupid hackish thing I was doing when grapping the image
buffer storing the last render.  Instead of getting it by allocating
the buffer again, store the pointer in a different state variable.  Now
you only call allocate when you need a new buffer.

With this change, you only call the state buffer allocate functions when
you want a fresh buffer.  In debug mode, fill the values with 0xDC to
make it more obvious when I create an error when reusing buffers.