Only allow splits before gathers in radix-k.
Previously when computing partition indices, the decision of split
versus gather was based on comparing the number of partitions to create
to the max number of splits.  In some circumstances, this could mean
a round gathers (and some processes drop out) and in a later round
splits happen.  This really complicates the indexing and was causing
errors.  Fix the problem by disallowing any splits after encountering
a round where splits should not take place.