Remove image_dest optimization from radix-k (mostly).
The point was that if a split did not happen, collect to the image_dest
if possible.  This way, the pixels would not have to be transferred
later.  However, in the real use case, there are not that many pixels
to transfer.  The use case of transfering a large amount of pixels
(such as a tree-like composite) does not currently exist.